board: qemu-arm: support receiving a transfer list
Adds support to receive a transfer list from previous boot stages if
configured with CONFIG_FIRMWARE_HANDOFF.

The two registers r1/x1 and r3/x3 are saved and later used when looking
for the FDT blob in board_fdt_blob_setup(). If a transfer list with a
FDT blob is found then that is used else we're falling back to the
default location as a way of being backwards compatible.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>